Update structure of API to follow MC
This a big update for the project. The structure of the API has to
changed to closer align with a Model-Controller (MC) pattern. As this
project has grown it has become harder to maintain the SQL queries for
the database so the decision was made to switch to using models to
represent the tables which removes the complexity of maintaining the
growing number of different interactions with the database.

Note: There are currently no plans to add Views (such as a MVC) pattern
to this project. At present it is a standalone API and front-end
applications will be developed in other technologies.